feat: keep old app folder
